requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4
supabase==2.3.4

# AI APIs
anthropic==0.18.1
openai==1.12.0
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.1
httpx==0.25.2
asyncio==3.4.3
aiohttp==3.9.3
email-validator==2.1.1

# Logging & Monitoring
loguru==0.7.2

# Date & Time
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3

# Invalidate cache
